the authorization page is divided into components
